add user to Day model, filter based on user, handle empty response

+++ server/controllers/dayController.js
@@ -12,7 +12,8 @@ const createDay = async (req, res) => {
   }
 
   const parsedDate = new Date(date);
-  const day = new Day({ date: parsedDate, score, tags });
+  // eslint-disable-next-line no-underscore-dangle
+  const day = new Day({ date: parsedDate, score, tags, user: req.user._id });
   const savedDay = await day.save();
 
   return res.status(201).json(savedDay);
@@ -25,6 +26,14 @@ const removeDay = async (req, res) => {
     return res.status(400).json({ error: 'Missing ID parameter' });
   }
 
+  const dayToDelete = await Day.findOne({ _id: id });
+  // eslint-disable-next-line no-underscore-dangle
+  if (dayToDelete.user.toString() !== req.user._id.toString()) {
+    return res
+      .status(401)
+      .json({ message: 'Cannot delete day of another user' });
+  }
+
   try {
     await Day.deleteOne({ _id: id });
     return res.status(200).json({ message: 'Deletion successful' });
@@ -35,7 +44,8 @@ const removeDay = async (req, res) => {
 
 const getDays = async (req, res) => {
   try {
-    const days = await Day.find({});
+    // eslint-disable-next-line no-underscore-dangle
+    const days = await Day.find({ user: req.user._id });
     res.json(days);
   } catch (error) {
     res.status(500).json({ error: 'Internal server error' });
